use armbox.conf instead of make.env
Idea is that rather than requiring the caller of compile.sh to set
$BOARD, compile.sh itself sources (indirectly via armbox-config) the
user-supplied env file that must set $BOARD. This makes it easier to
ensure that the settings from the file are set in scenaraios like
compile.sh docker, which it's not so easy to pass the caller's
environment down to the callee.

The file must be created as ./armbox.conf.