docker builds: DOCKER_HOST_SRC configurable
In cases where docker runs (./compile.sh docker ...) are run from inside
a docker container themselves, you can now set a variable
DOCKER_HOST_SRC to the path that $SRC has on the host. This might be
needed because the docker run needs to mount $SRC into the build
container, usually as -v $SRC/:/root/armbian/. $SRC in this case
is resolved on the docker host, and it might be a different path there.
That's why we're now using $DOCKER_HOST_SRC for this, if set. The
default value is $SRC.

Signed-off-by: Olaf Klischat <olaf.klischat@gmail.com>